requirements-dev.txt
# Runtime dependencies
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-benchmark>=3.4.0
memory-profiler>=0.60.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Security integrations (optional)
Flask-Login>=0.5.0
Flask-Security-Too>=4.0.0
Flask-JWT-Extended>=4.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
